Add and reposition pins on map
With a long press add a pin to the map at the location selected. If the the user keeps their finger on the pin they can reposition the pin. Once the user removes their finger from the screen the pin is dropped and cannot be moved -- at this time. When the pin is placed the location is geocoded and a placename is displayed.