Add the ability to remove pins
This commit adds the ability to remove pins from the map by tapping on a pin. To accomodate this the boolean isEditingMap was added to track editing state. To enter an editing state the user taps the edit button. The same button can be tapped to end editing. The UI is modified to visually relate the current editing state to the user.