Add Client for Downloading Photos from Flickr
This commit adds the client for downloading photos from Flickr for a given location. When a pin is dropped on the map image data is downloaded for that location if available. At this point nothing is done with the data, but utlimately it will be stored in a CoreData model and displayed on the GalleryViewController.