Download and display photos [in-progress]
Downloads images from the flickr API for a given location and displays them in the GalleryViewController.

The download is initiated at pin drop and therefore not necessarily complete by the time the user taps on the pin to display the gallery. If this happens, the app does not crash but will not download images until returning to the map view and the gallery view will be blank. This will be fixed in a later commit.

Also in progress is the ability to delete photos. I have added a button that will ultimately by used to delete single photos and the new collection button will delete all photos and download a new collection from Flickr.