Delete photos
This commits adds the ability to delete photos from a gallery. If the user taps on the edit button favorite icons are displayed on each of the gallery images. By tapping an image the favorite icon turns into an outline of a heart. The user can tap as many images as they wish to remove and then tap the "Update Favorites" button.

Tapping the Update Favorites button iterates through an array of imagesToDelete and proceeds to delete each image from Core Data. The work to update the collection view is handled in the NSFetchedResultsControllerDelegate functions where the index paths for each image are added to an array and those indexPaths are removed from the collection view.