Add ability to download all new gallery photos
Tapping on the "New Gallery" button will delete all existing photos associated with the location and download new images and display them in the collection view. Presently a blank cell is not displayed while the image is being downloaded. This is likely due to the batch downloader for images which I plan to change in a later commit.