Download new gallery [in-progress]
This commit is a start at separating the download of the image meta data from the actual images. This allows for the images to be downloaded individually rather than in a batch. This works at a basic level but is not quite there... the page is randomized between 1 and 1000 but often when selecting a remote location zero images are returned. This will need to be fixed.