Refactor and document Client class and associated functions
Refactors and documents the Client class and its associated functions. One of the goals was to eliminate dependencies and knowledge of the MapKit, CoreLocation and CoreData frameworks so that the Client model can stand on its own. This required a substantial rewrite of how the data is handled in the views, which will need to be tweaked further.