Change DownloadStatusView into a subclass of UIView
I created the DownloadStatusView as a subclass of UIView so that it's display and functionality is self-contained. The code will need to be refactored for conciseness but this is a good state to save my progress thus far. To accompany the DownloadStatusView subclass is a protocol for when the button is tapped by the user.