Change Core Data model
This commit changes the core data model to include a totalPages property to the pin and changes the value types for the various properties, to simplify use of the values.

The totalPages property will be used for fetching a random page of photos from Flickr. The idea is that by using the totalPages as an upper bound when generating a random integer the page retrieved will always fall within the accessible limits. As it stands, calling a page number greater than the total number of pages for a search will sometimes return results, but this results in erratic behavior with many locations not returning photos even if photos exist.