Randomize page number within range
When the initial gallery for a location has been downloaded, which is always page 1, the total number of pages is stored in the pin for that location. This change was implemented in commit 034878315c960db9c700a89dd5712ac06b5308d5. This commit takes the total page value (let's call this n) and generates a random number between 1 and n. This randomized value is input into the search query to, potentially, download different results. I say potentially because I have noticed when querying the Flickr API that it will sometimes return very similar, if not identical, results for each page. This is not a behavior of the app but of the API itself as I recgnized this behavior using Postman as well as with Flickr's API Explorer tool.

This commit also adds two new query keys/values to adjust the accuracy of the results: accuracy and radius.