FIX: Status view automatically shows complete on first download
Because both downloaded and total were initialized with a value of 0, when the values are compared for the very first time the application would set the status to complete despite picture data not having been downloaded. By changing the value of total to a non-zero value the application will now show the correct status in the Status View, even on first comparison.