FIX: Edit button enabled even with zero pins
This commit fixes a problem which had the edit button enabled even with zero pins on the map. The cause was that the button was enabled when a pin was dropped but if the pin was removed because no photos were available or the location was unknown the state was not updated. This commit fixes that problem by adding an ignoreEditState parameter to remove(annotation:from:) and replacing calls to DataController.delete(pin:with:from:completion:) with remove(annotation:from:ignoreEditState:).