Add EmptyGalleryView
Adds a view that is displayed when the gallery is empty. This view will display a button to download a new gallery if the user is not actively editing the gallery, i.e., the user removes all of the photos, exits to the map view and then  returns to the gallery by tapping on the pin. This is intended to reduce redunancy on the screen by only shoing one button to generate a new gallery.