FIX: Unnecessary call to DownloadStatusView.setVisible(:animated:)
When a pin is dropped in a location with no photos, such as the middle of the ocean the DownloadStatusView.setVisible(:animated:) method would be called a a second time which causes the view to animate to a visible position twice in quick succession and appears as a quick blink. This is not as noticable in the simulator but is much more apparent on an actual device. This commit removes this second call and appears to fix the problem.