FIX: Fatal error when pin is removed and user is touching the screen
This commit fixes a fatal runtime error wherein if a pin is automatically removed because it was dropped in an unknown location, or a location with no photos, and the user is actively touching the screen the application would crash because the value of activePin had been set to nil. This is fixed by checking for nil before trying to reposition the pin.